Split controller into 4 controllers
