+++ sdc.py
@@ -1,13 +1,13 @@
 import sys
 from PySide2.QtWidgets import QApplication
-from src.controller import Controller
+from src.Controllers.main_controller import MainController
 from src.webserver import flask_app
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
 
     flask = flask_app.FlaskThread()
-    ctrl = Controller()
+    ctrl = MainController()
     flask.init_controller(ctrl)
     ctrl.gui.show()
